added basic override function
Overrides the normal daily map cycle with specified override.txt/cfg upon the dc_override cvar being set from default of 0 -> 1. This would be useful at least for emergencies when needing to override the dailychanger plugin to a more vanilla config/mapcycle - without removing the plugin manually.

At most it would be great for setting different pre-set configs to run over the dailychanger for certain events or immediate server changes needed to be run over a set number of days.